VFB-2 EXtract modal into separate file
