+++ package.json
@@ -69,7 +69,6 @@
     "nyc": "^15.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-modal": "^3.16.1",
     "start-server-and-test": "^2.0.0",
     "styled-components": "^6.0.0-rc.3",
     "supabase": "^1.68.6",
